+++ models/Metrics/controller.js
@@ -1,5 +1,5 @@
-import { getSubscriptions, getSpecifications, getDefinitions, getInsights } from "../../../libs"
-import { parseSubscriptions, parseSpecifications, parseDefinitions } from "../../../utils"
+import { getSubscriptions, getSpecifications, getDefinitions, getInsights } from "../../libs"
+import { parseSubscriptions, parseSpecifications, parseDefinitions } from "../../utils"
 
 /* 
 Request and process data for metrics. The "sync" prefix indicates that these methods
